common code for rgb layers
