+++ keyboards/ergohaven/imperial44/keymaps/v1/keymap.c
@@ -3,25 +3,25 @@
 #include "oled/ergohaven_dark.c"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-        [_BASE] = LAYOUT( 
-          KC_ESC,  KC_Q,   KC_W,    KC_E,    KC_R,    KC_T,                                          KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,     KC_BSPC,  
-          KC_TAB,  KC_A,   KC_S,    KC_D,    KC_F,    KC_G,                                          KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN,  KC_RGUI,  
-          KC_LSFT, KC_Z,   KC_X,    KC_C,    KC_V,    KC_B,                                          KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_QUOT,  KC_RSFT,  
-                                             KC_LCTL, LOWER,   KC_SPC, KC_MUTE,    KC_HOME, KC_ENT,  RAISE,   KC_RALT 
+        [_BASE] = LAYOUT(
+          KC_ESC,  KC_Q,   KC_W,    KC_E,    KC_R,    KC_T,                                          KC_Y,    KC_U,    KC_I,    KC_O,    KC_P,     KC_BSPC,
+          KC_TAB,  KC_A,   KC_S,    KC_D,    KC_F,    KC_G,                                          KC_H,    KC_J,    KC_K,    KC_L,    KC_SCLN,  KC_RGUI,
+          KC_LSFT, KC_Z,   KC_X,    KC_C,    KC_V,    KC_B,                                          KC_N,    KC_M,    KC_COMM, KC_DOT,  KC_QUOT,  KC_RSFT,
+                                             KC_LCTL, LOWER,   KC_SPC, KC_MUTE,    KC_HOME, KC_ENT,  RAISE,   KC_RALT
         ),
 
-        [_LOWER] = LAYOUT( 
+        [_LOWER] = LAYOUT(
          _______, KC_1,    KC_2,    KC_3,    KC_4,    KC_5,                                          KC_6,    KC_7,    KC_8,    KC_9,    KC_0,     _______,
          _______, KC_HOME, KC_INS,  KC_DEL,  KC_END,  LANG,                                          _______, KC_LEFT, KC_DOWN, KC_UP,   KC_RIGHT, _______,
          _______, KC_UNDO, KC_CUT,  KC_COPY, KC_PSTE, KC_PSCR,                                       KC_CAPS, PREVWRD, KC_PGDN, KC_PGUP, NEXTWRD,  _______,
-                                             _______, _______, _______, _______,   _______, _______, ADJUST,  _______ 
+                                             _______, _______, _______, _______,   _______, _______, ADJUST,  _______
         ),
 
-        [_RAISE] = LAYOUT( 
+        [_RAISE] = LAYOUT(
          _______, KC_AT,   KC_LT,   KC_EQL,  KC_GT,   KC_GRV,                                        KC_CIRC, KC_LBRC, KC_UNDS, KC_RBRC, _______,  _______,
          _______, KC_BSLS, KC_LPRN, KC_MINS, KC_RPRN, KC_PLUS,                                       KC_PERC, KC_LCBR, KC_SCLN, KC_RCBR, KC_EXLM,  _______,
          _______, KC_HASH, KC_ASTR, KC_COLN, KC_SLSH, _______,                                       _______, KC_PIPE, KC_TILD, KC_AMPR, KC_DLR,   _______,
-                                             _______, ADJUST,  _______, _______,   _______, _______, _______, _______ 
+                                             _______, ADJUST,  _______, _______,   _______, _______, _______, _______
         ),
 
       [_ADJUST] = LAYOUT(
@@ -141,46 +141,4 @@ bool oled_task_user(void) {
     return false;
 }
 
-#endif
-
-#ifdef RGBLIGHT_ENABLE
-const rgblight_segment_t PROGMEM my_base_layer[] = RGBLIGHT_LAYER_SEGMENTS(
-    {0, 4, HSV_WHITE}
-);
-const rgblight_segment_t PROGMEM my_layer1_layer[] = RGBLIGHT_LAYER_SEGMENTS(
-    {0, 4, HSV_CYAN}
-);
-const rgblight_segment_t PROGMEM my_layer2_layer[] = RGBLIGHT_LAYER_SEGMENTS(
-    {0, 4, HSV_RED}
-);
-const rgblight_segment_t PROGMEM my_layer3_layer[] = RGBLIGHT_LAYER_SEGMENTS(
-    {0, 4, HSV_GOLDENROD}
-);
-
-// Now define the array of layers. Later layers take precedence
-const rgblight_segment_t* const PROGMEM my_rgb_layers[] = RGBLIGHT_LAYERS_LIST(
-    my_base_layer,
-    my_layer1_layer,
-    my_layer2_layer,
-    my_layer3_layer
-);
-
-
-void keyboard_post_init_user(void) {
-    // Enable the LED layers
-    rgblight_layers = my_rgb_layers;
-}
-
-layer_state_t default_layer_state_set_user(layer_state_t state) {
-    rgblight_set_layer_state(0, layer_state_cmp(state, _BASE));
-    return state;
-}
-
-layer_state_t layer_state_set_user(layer_state_t state) {
-    rgblight_set_layer_state(1, layer_state_cmp(state, _RAISE));
-    rgblight_set_layer_state(2, layer_state_cmp(state, _LOWER));
-    rgblight_set_layer_state(3, layer_state_cmp(state, _ADJUST));
-    return state;
-
-}
 #endif
