+++ keyboards/ergohaven/k03/rules.mk
@@ -10,3 +10,5 @@ OLED_DRIVER = ssd1306
 SPLIT_KEYBOARD = yes
 SERIAL_DRIVER = vendor
 RGBLIGHT_ENABLE = yes
+
+SRC += keyboards/ergohaven/ergohaven_rgb.c
