+++ keyboards/ergohaven/planeta/rules.mk
@@ -8,3 +8,5 @@ DEFAULT_FOLDER = ergohaven/planeta/rev2
 
 AUDIO_ENABLE = yes
 AUDIO_DRIVER = pwm_hardware
+
+SRC += keyboards/ergohaven/ergohaven_rgb.c
